refactor: menus 패키지 분리
