+++ src/main/java/kitchenpos/menus/domain/JpaMenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menus.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
