+++ src/main/java/kitchenpos/menus/domain/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menus.domain;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
