+++ src/main/java/kitchenpos/menus/domain/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menus.domain;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
