+++ src/main/java/kitchenpos/menus/domain/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menus.domain;
 
 import javax.persistence.*;
 import java.util.UUID;
