+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -1,7 +1,11 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.*;
 import kitchenpos.common.domain.PurgomalumClient;
+import kitchenpos.menus.application.MenuService;
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuGroupRepository;
+import kitchenpos.menus.domain.MenuProduct;
+import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
